+++ handlers/custom.js
@@ -1,5 +1,4 @@
-var Dota2 = require("../index"),
-    util = require("util");
+var Dota2 = require("../index");
 
 /**
  * Requests a list of custom game modes for which there are currently lobbies available.
@@ -12,11 +11,10 @@ Dota2.Dota2Client.prototype.requestJoinableCustomGameModes = function requestJoi
     server_region = server_region || Dota2.ServerRegion.UNSPECIFIED;
     
     this.Logger.debug("Sending joinable custom game modes request");
-    var payload = {
+    var payload = new Dota2.schema.CMsgJoinableCustomGameModesRequest({
         "server_region": server_region
-    };
-    this.sendToGC(  Dota2.schema.lookupEnum("EDOTAGCMsg").values.k_EMsgGCJoinableCustomGameModesRequest, 
-                    Dota2.schema.lookupType("CMsgJoinableCustomGameModesRequest").encode(payload).finish());
+    });
+    this.sendToGC(Dota2.schema.EDOTAGCMsg.k_EMsgGCJoinableCustomGameModesRequest, payload);
 }
 
 // Events
@@ -30,9 +28,9 @@ Dota2.Dota2Client.prototype.requestJoinableCustomGameModes = function requestJoi
 var handlers = Dota2.Dota2Client.prototype._handlers;
 
 var onJoinableCustomGameModesResponse = function onJoinableCustomGameModesResponse(message) {
-    var modes = Dota2.schema.lookupType("CMsgJoinableCustomGameModesResponse").decode(message);
+    var modes = Dota2.schema.CMsgJoinableCustomGameModesResponse.decode(message);
     
     this.Logger.debug("Received joinable custom game modes");
     this.emit("joinableCustomGameModes", modes.game_modes);
 };
-handlers[Dota2.schema.lookupEnum("EDOTAGCMsg").values.k_EMsgGCJoinableCustomGameModesResponse] = onJoinableCustomGameModesResponse;
+handlers[Dota2.schema.EDOTAGCMsg.k_EMsgGCJoinableCustomGameModesResponse] = onJoinableCustomGameModesResponse;